backend/requirements.txt
fastapi
uvicorn
onnxruntime
pydantic
python-multipart
tensorflow
pillow